finished add functions api requests

+++ pages/api/products/[id].js
@@ -5,7 +5,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
 
   dbConnect();
 
@@ -19,6 +21,10 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Error: User not authorized.");
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -29,6 +35,10 @@ export default async function handler(req, res) {
     }
   }
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Error: User not authorized.");
+    }
+
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted");
